Split test for hip_intra_block_reduce_scan into a separate file
