+++ core/unit_test/TestRange.hpp
@@ -433,59 +433,8 @@ TEST(TEST_CATEGORY, range_reduce) {
   }
 }
 
-#ifdef KOKKOS_ENABLE_HIP
-struct DummyFunctor {
-  using value_type = int;
-  void operator()(const int, value_type &, bool) const {}
-};
-
-template <int N>
-__global__ void start_intra_block_scan() {
-  __shared__ DummyFunctor::value_type values[N];
-  const int i = hipThreadIdx_y;
-  values[i]   = i + 1;
-  __syncthreads();
-
-  DummyFunctor f;
-  Kokkos::Impl::hip_intra_block_reduce_scan<true, DummyFunctor, void>(f,
-                                                                      values);
-
-  __syncthreads();
-  if (values[i] != ((i + 2) * (i + 1)) / 2) {
-    printf("Value for %d should be %d but is %d\n", i, ((i + 2) * (i + 1)) / 2,
-           values[i]);
-    Kokkos::abort("Test for intra_block_reduce_scan failed!");
-  }
-}
-
-template <int N>
-void test_intra_block_scan() {
-  dim3 grid(1, 1, 1);
-  dim3 block(1, N, 1);
-  hipLaunchKernelGGL(start_intra_block_scan<N>, grid, block, 0, 0);
-}
-#endif
-
 #ifndef KOKKOS_ENABLE_OPENMPTARGET
 TEST(TEST_CATEGORY, range_scan) {
-#ifdef KOKKOS_ENABLE_HIP
-  if (std::is_same<
-          TEST_EXECSPACE,
-          typename Kokkos::Experimental::HIPSpace::execution_space>::value) {
-    test_intra_block_scan<1>();
-    test_intra_block_scan<2>();
-    test_intra_block_scan<4>();
-    test_intra_block_scan<8>();
-    test_intra_block_scan<16>();
-    test_intra_block_scan<32>();
-    test_intra_block_scan<64>();
-    test_intra_block_scan<128>();
-    test_intra_block_scan<256>();
-    // FIXME_HIP block sizes larger than 256 give wrong results.
-    // test_intra_block_scan<512>();
-    // test_intra_block_scan<1024>();
-  }
-#endif
   {
     TestRange<TEST_EXECSPACE, Kokkos::Schedule<Kokkos::Static> > f(0);
     f.test_scan();
